Refactor Sidebar, add Collapsible, update dependencies
