+++ app/dashboard/[workspaceId]/application/_components/Sidebar/index.tsx
@@ -1,113 +1,39 @@
-"use client";
-import { Input } from "@components/ui/input";
-import { useDraggable } from "@dnd-kit/core";
-import React, { useState } from "react";
+import React from "react";
+import CompSidebar from "./Component";
 
-type SidebarProps = {
-  children?: React.ReactNode;
-};
+type Props = {};
 
-const test = [
-  {
-    id: 11,
-    name: "Dashboard",
-    icon: "📊",
-    description: "View and manage your analytics and reports.",
-  },
-  {
-    id: 21,
-    name: "CRM",
-    icon: "📇",
-    description: "Manage customer relationships and interactions.",
-  },
-  {
-    id: 31,
-    name: "Project Management",
-    icon: "📅",
-    description: "Organize and track your projects and tasks.",
-  },
-  {
-    id: 41,
-    name: "Marketing",
-    icon: "📢",
-    description: "Plan and execute marketing campaigns.",
-  },
-  {
-    id: 51,
-    name: "Dashboard",
-    icon: "📊",
-    description: "View and manage your analytics and reports.",
-  },
-  {
-    id: 61,
-    name: "CRM",
-    icon: "📇",
-    description: "Manage customer relationships and interactions.",
-  },
-  {
-    id: 71,
-    name: "Project Management",
-    icon: "📅",
-    description: "Organize and track your projects and tasks.",
-  },
-  {
-    id: 81,
-    name: "Marketing",
-    icon: "📢",
-    description: "Plan and execute marketing campaigns.",
-  },
-];
+// Add no. of Sidebar as per need
+const Navigators = Array(3).fill(false);
 
-// Draggable Component
-const Draggable = ({ id, name, icon, description }: any) => {
-  const { attributes, listeners, setNodeRef, transform } = useDraggable({
-    id: id,
-    data: {
-      type: "item",
-    },
-  });
-  return (
-    <div
-      ref={setNodeRef}
-      style={{
-        transform: transform
-          ? `translate(${transform.x}px, ${transform.y}px)`
-          : undefined,
-      }}
-      {...attributes}
-      {...listeners}
-      className="bg-white/20 p-2 rounded-lg text-center"
-    >
-      {name}
-    </div>
-  );
-};
+const Sidebar = (props: Props) => {
+  const [State, setState] = React.useState(Navigators);
 
-const Sidebar = () => {
-  const { attributes, listeners, setNodeRef, transform } = useDraggable({
-    id: "sidebar",
-    data: {
-      type: "sidebar",
-    },
-  });
+  const handleClick = (index: number) => {
+    const newState = [...Navigators];
+    if (!State[index]) {
+      newState[index] = !newState[index];
+    }
+    setState(newState);
+    console.log(newState);
+  };
 
   return (
-    <div className="">
-      <div className="bg-blue-800 h-fill h-screen">
-        <div className="h-[100px] flex items-center justify-center">
-          <Input
-            className=" text-white bg-white/20 p-2 rounded-lg w-full "
-            placeholder="Search ..."
-          ></Input>
-        </div>
-        <div className="flex flex-col py-[10%]">
-          <div className="grid grid-cols-2 gap-5 mx-2">
-            {test.map((item) => (
-              <Draggable {...item} key={item.id} />
-            ))}
+    <div className="flex">
+      <div className="w-[100px] h-full bg-white  flex-col space-y-10">
+        {Navigators.map((item, index) => (
+          <div
+            key={index}
+            className="text-black text-center cursor-pointer bg-red-300 w-full h-12 flex items-center justify-center"
+            onClick={() => handleClick(index)}
+          >
+            Click Here
           </div>
-        </div>
+        ))}
       </div>
+      <CompSidebar open={State[0]} />
+      <CompSidebar open={State[1]} />
+      <CompSidebar open={State[2]} />
     </div>
   );
 };
